+++ src/entities/news/api/rssParser.ts
@@ -1,12 +1,11 @@
 import { parse } from 'rss-to-json';
-import type { NewsItem } from '../model';
+import type { NewsItem, RssFeedItem } from '../model';
 
-// TODO: AI로 생성한 보일러 플레이트 - 개선 필요
 export async function fetchRssFeed(feedUrl: string, sourceName: string): Promise<NewsItem[]> {
   try {
     const feed = await parse(feedUrl);
 
-    return feed.items.map((item: any) => {
+    return feed.items.map((item: RssFeedItem) => {
       // 썸네일 이미지 추출 (RSS 형식에 따라 다를 수 있음)
       let thumbnail = '';
