+++ src/entities/news/model/index.ts
@@ -1,9 +1,1 @@
-export interface NewsItem {
-  title: string;
-  link: string;
-  description?: string;
-  pubDate: string;
-  thumbnail?: string;
-  source: string;
-  guid: string;
-}
+export * from './types';
